+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Can register a setting alias directly on the setting
+
 ## [v1.0.2] - 2022-02-11
 
 ### Fixed